done with all basic api of crud

+++ delete.php
@@ -11,15 +11,22 @@ include "connection.php";
 $jon =json_decode(file_get_contents("php://input"),true);
 
 $od=$jon['sid'];
-$id  = "DELETE from `API`.`contacts` where `id` = {$od}";
-
+$getId="SELECT  * from `API`.`contacts` where `id` = {$od}"; //  for validation of id
+$result= mysqli_query($con, $getId) or die("query failed");
+    
+if(mysqli_num_rows($result)>0){    //  validation of id
+    $id  = "DELETE from `API`.`contacts` where `id` = {$od}"; // delete only if id exist
 
 if(mysqli_query($con, $id)) {
     echo json_encode(array('message' => 'Record Deleted Successfully', 'status'=>'true'));
  }
- else{
-   echo json_encode(array('message' => 'Not Successful', 'status'=>'false'));
- }
+ 
+
+}
+else{
+    echo json_encode(array('message' => 'ID doesnot Exist', 'status'=>'false'));
+  }
+
